+++ OnlineExamBe/OnlineExam/Controllers/CLassController.cs
@@ -1,27 +1,33 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using OnlineExam.Application.Dtos.ClassDtos;
 using OnlineExam.Application.Dtos.RequestDtos.UserDtos;
 using OnlineExam.Application.Dtos.ResponseDtos;
 using OnlineExam.Application.Interfaces;
 using OnlineExam.Attributes;
 using OnlineExam.Domain.Enums;
+using OnlineExam.Infrastructure.Policy.Requirements;
 using System.Text.Json;
 
 namespace OnlineExam.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
+    
     public class CLassController : Controller
     {
         private readonly IClassService _classService;
-        public CLassController(IClassService classService) 
+        private readonly IAuthorizationService _authorizationService;
+        public CLassController(IClassService classService, IAuthorizationService authorizationService) 
         {
             _classService = classService;
+            _authorizationService = authorizationService;
 
         }
+        //admin
         [HttpGet]
         [Route("get-all")]
-        [SessionAuthorize(UserRole.ADMIN)]
+        [SessionAuthorize]
         public async Task<IActionResult> GetAll()
         {
             ResultApiModel apiResultModel = new ResultApiModel();
@@ -32,22 +38,39 @@ namespace OnlineExam.Controllers
         }
         [HttpGet]
         [Route("get-by-teacher-and-subject")]
-        [SessionAuthorize(UserRole.TEACHER,UserRole.ADMIN)]
+        [SessionAuthorize("F0112")]
         public async Task<IActionResult> GetByTeacherAndSubject(int? teacherId = null, int? subjectId = null )
         {
+            
             ResultApiModel apiResultModel = new ResultApiModel();
             apiResultModel = await _classService.GetByTeacherAndSubject(teacherId, subjectId);
             return Ok(apiResultModel);
         }
 
-
+        [HttpGet]
+        [Route("get-by-id/{classId}")]
+        [SessionAuthorize("F0112")]
+        public async Task<IActionResult> GetById(int classId)
+        {
+            ResultApiModel apiResultModel = new ResultApiModel();
+            var c = await _classService.GetByIdAsync(classId, ["Teacher", "Subject", "Exams", "StudentClasses"]);
+            var authResult = await _authorizationService.AuthorizeAsync(User, c, new ResourceRequirement(ResourceAction.ViewDetail));
+            if (!authResult.Succeeded)
+            {
+                return Unauthorized("Forbidden: You do not have permission to perform this action.");
+            }
+            if (c != null) apiResultModel.Status = true;
+            apiResultModel.Data = new ClassDto(c);
+            return Ok(apiResultModel);
+        }
         [HttpGet]
         [Route("get-students")]
-        
+        [SessionAuthorize("F0000")]
         public async Task<IActionResult> GetStudents(int classId)
         {
             ResultApiModel apiResultModel = new ResultApiModel();
             apiResultModel = await _classService.GetStudents(classId);
+            if (apiResultModel.MessageCode == ResponseCode.Forbidden) return Unauthorized("Forbidden: You do not have permission to perform this action.");
             return Ok(apiResultModel);
         }
         /// <summary>
@@ -58,8 +81,8 @@ namespace OnlineExam.Controllers
         /// <returns></returns>
         [HttpPost]
         [Route("add-users/{classId}")]
-        //[Authorize(Roles ="ADMIN")]
-        public async Task<IActionResult> AddUsers(IFormFile file,int classId)
+        [SessionAuthorize("F0113")]
+        public async Task<IActionResult> AddStudents(IFormFile file,int classId)
         {
             if (file == null || file.Length == 0)
                 return BadRequest("File is empty");
@@ -74,13 +97,27 @@ namespace OnlineExam.Controllers
                 return BadRequest("Invalid JSON file");
 
             var result = await _classService.AddStudentsAsync(listUser, classId);
+            if (result.MessageCode == ResponseCode.Forbidden) return Unauthorized("Forbidden: You do not have permission to perform this action.");
+            return Ok(result);
+        }
 
+        [HttpPost]
+        [Route("add-user/{classId}")]
+        [SessionAuthorize("F0113")]
+        public async Task<IActionResult> AddStudent(AddStudentDto student, int classId)
+        {
+            
+            if (student == null)
+                return BadRequest("Invalid JSON file");
+
+            var result = await _classService.AddStudentAsync(student, classId);
+            if (result.MessageCode == ResponseCode.Forbidden) return Unauthorized("Forbidden: You do not have permission to perform this action.");
             return Ok(result);
         }
 
         [HttpPost]
         [Route("create")]
-        //[Authorize(Roles ="ADMIN")]
+        [SessionAuthorize("F0101")]
         public async Task<IActionResult> Create(CreateClassDto newClass)
         {
             ResultApiModel apiResultModel = new ResultApiModel();
@@ -91,11 +128,12 @@ namespace OnlineExam.Controllers
        
         [HttpPut]
         [Route("update/{classId}")]
-
+        [SessionAuthorize("F0113")]
         public async Task<IActionResult> Update(UpdateClassDto updateClass, int classId)
         {
             ResultApiModel apiResultModel = new ResultApiModel();
             apiResultModel = await _classService.UpdateAsync(updateClass, classId);
+            if (apiResultModel.MessageCode == ResponseCode.Forbidden) return Unauthorized("Forbidden: You do not have permission to perform this action.");
             return Ok(apiResultModel);
         }
 
@@ -103,10 +141,16 @@ namespace OnlineExam.Controllers
        
         [HttpDelete]
         [Route("delete")]
-
+        [SessionAuthorize("F0114")]
         public async Task<IActionResult> Delete(int classId)
         {
             ResultApiModel apiResultModel = new ResultApiModel();
+            var curClass = await _classService.GetByIdAsync(classId);
+            var authResult = await _authorizationService.AuthorizeAsync(User, curClass, new ResourceRequirement(ResourceAction.ViewDetail));
+            if (!authResult.Succeeded)
+            {
+                return Unauthorized("Forbidden: You do not have permission to perform this action.");
+            }
             var success = await _classService.DeleteAsync(classId);
             apiResultModel.Status = success;
             apiResultModel.Data = success;
